requirements.txt
python-telegram-bot[webhooks]
requests
fastapi
gunicorn
uvicorn
playwright